# openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Responder",
    "version": "1.0.0"
  },
  "paths": {
    "/responder": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Responder"
              },
              "examples": {
                "Create Responder Example": {
                  "value": {
                    "name": "John Doe",
                    "phoneNumber": "111-222-333",
                    "latitude": 30.12345,
                    "longitude": -70.98765,
                    "boatCapacity": 3,
                    "medicalKit": true,
                    "available": true
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Success"
          }
        }
      }
    },
    "/responders/available": {
      "get": {
        "summary": "Get List of Available Responders",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponderList"
                },
                "examples": {
                  "List  Of Responders": {
                    "value": [
                      {
                        "id": "1",
                        "name": "John Doe",
                        "phoneNumber": "111-222-333",
                        "latitude": 30.12345,
                        "longitude": -70.98765,
                        "boatCapacity": 3,
                        "medicalKit": true,
                        "available": true
                      },
                      {
                        "id": "2",
                        "name": "Mary Foo",
                        "phoneNumber": "999-888-777",
                        "latitude": 32.989765,
                        "longitude": -71.12345,
                        "boatCapacity": 2,
                        "medicalKit": false,
                        "available": true
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/responder/byname/{name}": {
      "get": {
        "summary": "Get Responder By Name",
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Responder"
                },
                "examples": {
                  "Responder": {
                    "value": {
                      "id": "1",
                      "name": "John Doe",
                      "phoneNumber": "111-222-333",
                      "latitude": 30.12345,
                      "longitude": -70.98765,
                      "boatCapacity": 3,
                      "medicalKit": true,
                      "available": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Responder Not Found"
          }
        }
      },
      "parameters": [
        {
          "name": "name",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/responder/{id}": {
      "get": {
        "summary": "Get Responder by Id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Responder Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Responder"
                },
                "examples": {
                  "Responder": {
                    "value": {
                      "id": "1",
                      "name": "John Doe",
                      "phoneNumber": "111-222-333",
                      "latitude": 30.12345,
                      "longitude": -70.98765,
                      "boatCapacity": 3,
                      "medicalKit": true,
                      "available": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Responder": {
        "title": "Root Type for Responder",
        "description": "The root of the Responder type's schema.",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "latitude": {
            "format": "double",
            "type": "number"
          },
          "longitude": {
            "format": "double",
            "type": "number"
          },
          "boatCapacity": {
            "format": "int32",
            "type": "integer"
          },
          "medicalKit": {
            "type": "boolean"
          },
          "available": {
            "type": "boolean"
          }
        },
        "example": "{\n  \"id\" : \"1\",\n  \"name\": \"John Doe\",\n  \"phoneNumber\": \"111-222-333\",\n  \"latitude\": 30.12345,\n  \"longitude\": -70.98765,\n  \"boatCapacity\": 3,\n  \"medicalKit\": true,\n  \"available\": true\n}"
      },
      "ResponderList": {
        "title": "Root Type for ResponderList",
        "description": "The root of the ResponderList type's schema.",
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
            },
            "name": {
              "type": "string"
            },
            "phoneNumber": {
              "type": "string"
            },
            "latitude": {
              "format": "double",
              "type": "number"
            },
            "longitude": {
              "format": "double",
              "type": "number"
            },
            "boatCapacity": {
              "format": "int32",
              "type": "integer"
            },
            "medicalKit": {
              "type": "boolean"
            },
            "available": {
              "type": "boolean"
            }
          }
        },
        "example": [
          {
            "id": "1",
            "name": "John Doe",
            "phoneNumber": "111-222-333",
            "latitude": 30.12345,
            "longitude": -70.98765,
            "boatCapacity": 3,
            "medicalKit": true,
            "available": true
          },
          {
            "id": "2",
            "name": "Mary Foo",
            "phoneNumber": "999-888-777",
            "latitude": 32.989765,
            "longitude": -71.12345,
            "boatCapacity": 2,
            "medicalKit": false,
            "available": true
          }
        ]
      }
    }
  }
}